Move migration script into node.py
