+++ golem/node.py
@@ -49,6 +49,7 @@ from golem.rpc.session import (
 from golem import terms
 from golem.tools.uploadcontroller import UploadController
 from golem.tools.remotefs import RemoteFS
+from golem.task.acl import migrate_txt
 
 if TYPE_CHECKING:
     # pylint:disable=unused-import
@@ -135,6 +136,9 @@ class Node(HardwarePresetsMixin):
         # Initialize database
         self._db = Database(
             db, fields=DB_FIELDS, models=DB_MODELS, db_dir=datadir)
+
+        # ACL txt migration
+        migrate_txt(Path(datadir))
 
         self.client: Optional[Client] = None
