+++ golem/task/acl.py
@@ -4,7 +4,8 @@ import operator
 import time
 import ipaddress
 from enum import Enum
-from typing import Dict, Union, Iterable, Optional, Tuple, List, cast
+from pathlib import Path
+from typing import Dict, Set, Union, Iterable, Optional, Tuple, List, cast
 from sortedcontainers import SortedList
 from golem.model import ACLAllowedNodes, ACLDeniedNodes, GenericKeyValue
 
@@ -364,3 +365,33 @@ def _get_node_info(client, key: str) -> Dict:
         pass
 
     return node
+
+def migrate_txt(datadir):
+
+    def _read_set_from_file(path: Path) -> Set[str]:
+        try:
+            with path.open() as f:
+                return set(line.strip() for line in f)
+        except OSError:
+            return set()
+
+    def _remove_file(path: Path) -> None:
+        if(path.exists()):
+            path.unlink()
+
+    DENY_LIST_NAME = "deny.txt"
+    ALL_EXCEPT_ALLOWED = "ALL_EXCEPT_ALLOWED"
+    nodes_ids = []
+
+    deny_list_path = datadir / DENY_LIST_NAME
+    nodes_ids = _read_set_from_file(deny_list_path)
+    if(len(nodes_ids) > 0):
+        if ALL_EXCEPT_ALLOWED in nodes_ids:
+            nodes_ids.remove(ALL_EXCEPT_ALLOWED)
+            nodes = [{'node_id': node_id, 'node_name': None}
+                     for node_id in nodes_ids]
+            ACLAllowedNodes.insert_many(nodes).execute()
+        nodes = [{'node_id': node_id, 'node_name': None}
+                 for node_id in nodes_ids]
+        ACLDeniedNodes.insert_many(nodes).execute()
+    _remove_file(deny_list_path)
